backend_django/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
alabaster==0.7.12
anytree==2.8.0
appdirs==1.4.4
argcomplete==1.12.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
Beaker==1.11.0
beautifulsoup4==4.10.0
btrfsutil==5.15
CacheControl==0.12.6
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.3
cocos2d==0.6.9
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==35.0.0
Cython==0.29.24
defusedxml==0.7.1
distlib==0.3.3
distro==1.6.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docutils==0.17.1
elasticsearch==7.14.1
fake-useragent==0.1.11
future==0.18.2
geographiclib==1.52
geopy==2.2.0
Glances==3.2.3
glfw==2.4.0
googletransx==2.4.2
html5lib==1.1
idna==3.3
imagesize==1.2.0
imgui==1.4.1
Jinja2==3.0.2
joblib==1.1.0
lensfun==0.3.95
lit==12.0.1.dev0
louis==3.19.0
lxml==4.6.3
Mako==1.1.5
mallard-ducktype==1.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.2
multidict==5.1.0
mutagen==1.45.1
netsnmp-python==1.0a1
nltk==3.6.5
numpy==1.21.2
ordered-set==4.0.2
packaging==20.9
pandas==1.3.2
pep517==0.11.0
Pillow==8.3.2
ply==3.11
progress==1.6
psutil==5.8.0
pybind11==2.8.1
pycares==4.0.0
pycparser==2.20
pycryptodomex==3.11.0
pygame==2.1.0
pyglet==1.5.21
Pygments==2.10.0
PyGObject==3.42.0
PyJWT==2.3.0
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySDL2==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==1.2.4
pytz==2021.3
PyYAML==5.4.1
Reflector==2021.7.8.19.52.52
regex==2021.11.10
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
schedule==1.1.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
websockets==10.0
xmltodict==0.12.0
yarl==1.6.3
yq==2.12.2
yt-dlp==2021.10.10
